[polling-controller] minor type refactors (#4042)
## Explanation

- Consolidate types into 'types' module.
- Add comment explaining that `any` is required by the compiler for
mixin constructors.

## References

- Closes https://github.com/MetaMask/core/issues/3728

## Changelog

N/A

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
